Added test suite, finished test classes, added jar file. SHOULD be final version.

+++ NewSilentAuction/src/backend/Bid.java
@@ -40,7 +40,7 @@ public class Bid extends Contribution
 	 * 
 	 * @return the bidders id.
 	 */
-	public double getBidder()
+	public int getBidder()
 	{
 		return super.getUser();
 	}
